==> src/3_train_basemodel.ipynb <==
!nvidia-smi

!pip install transformers==4.20.1
!pip install seqeval==1.2.2
!pip install wandb==0.12.21
!pip install torchcontrib==0.0.2

!wandb disabled

from google.colab import drive
drive.mount('/content/drive')

# 自身の環境のパスを指定
base_folder = "drive/MyDrive/Colab\ Notebooks/cpt-hanrei-1st-refactor/src"

cd {base_folder}

for fold in range(5):
    for model_name in ["cl-wom", "cl-charwom", "cl","cl-char","NICT-100k","NICT-32k"]:
            !python run.py \
                --model_name={model_name} \
                --name="seed_data_for_each_model" \
                --epochs=1  \
                --fold={fold} \
                --warmup_ratio=0.1 \
                --train_batch_size=14
