==> src/6_infer.ipynb <==
!nvidia-smi

!pip install transformers==4.20.1
!pip install fugashi==1.1.2
!pip install ipadic==1.0.0
!pip install fsspec==2022.5.0
!pip install seqeval==1.2.2
!pip install mojimoji==0.0.12
!pip install mecab-python3==1.0.5

!apt install mecab-jumandic-utf8
!apt install mecab
!ln -s /etc/mecabrc /usr/local/etc/mecabrc

from google.colab import drive
drive.mount('/content/drive')

# 自身の環境のパスを指定
base_folder = "drive/MyDrive/Colab\ Notebooks/cpt-hanrei-1st-refactor/src"

cd {base_folder}

cd infer_basemodel

!python dataprocess.py

!python infer.py

cd ../infer_stacking

!python dataprocess.py

!python infer.py

stacking_infer_only.csvが提出ファイルとなる
